Move named expression operations into CrudOperations
