+++ src/HyperFormula.ts
@@ -4,7 +4,6 @@
  */
 
 import {AbsoluteCellRange} from './AbsoluteCellRange'
-import {absolutizeDependencies} from './absolutizeDependencies'
 import {CellType, CellValueType, getCellType, getCellValueType, NoErrorCellValue, SimpleCellAddress} from './Cell'
 import {CellContent, CellContentParser, RawCellContent} from './CellContentParser'
 import {CellValue, ExportedChange, Exporter} from './CellValue'
@@ -24,9 +23,6 @@ import {
 } from './DependencyGraph'
 import {
   EvaluationSuspendedError,
-  NamedExpressionDoesNotExist,
-  NamedExpressionNameIsAlreadyTaken,
-  NamedExpressionNameIsInvalid,
   NotAFormulaError
 } from './errors'
 import {Evaluator} from './Evaluator'
@@ -1423,15 +1419,7 @@ export class HyperFormula implements TypedEmitter {
    * @category Named Expression
    */
   public addNamedExpression(expressionName: string, expression: RawCellContent): ExportedChange[] {
-    if (!this._namedExpressions.isNameValid(expressionName)) {
-      throw new NamedExpressionNameIsInvalid(expressionName)
-    }
-    if (!this._namedExpressions.isNameAvailable(expressionName)) {
-      throw new NamedExpressionNameIsAlreadyTaken(expressionName)
-    }
-    const namedExpression = this._namedExpressions.addNamedExpression(expressionName)
-    const address = this._namedExpressions.getInternalNamedExpressionAddress(expressionName)!
-    this.storeExpressionInCell(address, expression)
+    this._crudOperations.addNamedExpression(expressionName, expression)
     const changes = this.recomputeIfDependencyGraphNeedsIt()
     this._emitter.emit(Events.NamedExpressionAdded, expressionName, changes)
     return changes
@@ -1487,11 +1475,7 @@ export class HyperFormula implements TypedEmitter {
    * @category Named Expression
    */
   public changeNamedExpression(expressionName: string, newExpression: RawCellContent): ExportedChange[] {
-    const address = this._namedExpressions.getInternalNamedExpressionAddress(expressionName)
-    if (!address) {
-      throw new NamedExpressionDoesNotExist(expressionName)
-    }
-    this.storeExpressionInCell(address, newExpression)
+    this._crudOperations.changeNamedExpressionExpression(expressionName, newExpression)
     return this.recomputeIfDependencyGraphNeedsIt()
   }
 
@@ -1683,20 +1667,6 @@ export class HyperFormula implements TypedEmitter {
       return exportedChanges
     } else {
       return []
-    }
-  }
-
-  private storeExpressionInCell(address: SimpleCellAddress, expression: RawCellContent) {
-    const parsedCellContent = this._cellContentParser.parse(expression)
-    if (parsedCellContent instanceof CellContent.MatrixFormula) {
-      throw new Error('Matrix formulas are not supported')
-    } else if (parsedCellContent instanceof CellContent.Formula) {
-      const {ast, hasVolatileFunction, hasStructuralChangeFunction, dependencies} = this._parser.parse(parsedCellContent.formula, address)
-      this._dependencyGraph.setFormulaToCell(address, ast, absolutizeDependencies(dependencies, address), hasVolatileFunction, hasStructuralChangeFunction)
-    } else if (parsedCellContent instanceof CellContent.Empty) {
-      this._crudOperations.operations.setCellEmpty(address)
-    } else {
-      this._crudOperations.operations.setValueToCell(parsedCellContent.value, address)
     }
   }
 }
